refactor(reconciliation): split server.go into domain endpoint files
Extract RPC handlers from server.go (999 lines) into domain-specific files:
- grpc_settlement_endpoints.go: Initiate, Retrieve, List runs and results
- grpc_pipeline_endpoints.go: Execute, Control, pipeline internals
- grpc_assertion_endpoints.go: AssertBalance and caller role extraction
- Rename dispute_handler.go -> grpc_dispute_endpoints.go
- Rename list_handler.go -> grpc_list_endpoints.go
- Move assertion mappers to mapping.go

server.go now contains only struct, interfaces, constructor, and options
(187 lines). Removed from knownOversizedFiles.
